# Implement precisely-allocated persistence