… apc_pool.c
… apc_pool.h
… apc_pool_api.h
… config.w32
… php_apc.c
… tests/apc_020.phpt
… tests/apc_023.phpt