Currently, persistence to SMA is pool based, i.e. SMA memory is allocated in blocks (of minimum size 512 bytes). This means that small entries (say entries storing integers) use much more memory than they need to and large entries will have a variable amount of additional overhead (depending on how many pool blocks are needed, how well they are filled and which block sizes are used).

This PR implementations a new mechanism, which performs persistence in two steps: First, we compute the amount of memory the persistence will require, then we perform a single SMA allocation of the correct size, and finally perform the actual copies into the allocated memory segment.

As such, this implementation only performs a single allocation of exactly the right size, at the cost of having to precompute the size in a separate pass. We also save 64 bytes in pool management overhead as a nice side benefit.

The implementation is not quite ready yet, some things left:

 * [x] This still uses the old code from copying from SHM. This can be simplified a lot now.
 * [x] Missing handling for INDIRECT zvals in arrays (can only occur in the unusal case of persisting the symbol table). There's currently no test for this.
 * [x] Missing test for nNextFreeElement handling. The previous implementation was buggy in this regard.
 * [x] Performance testing.