@nikic 2018-10-08
Some microbenchmark numbers using https://gist.github.com/nikic/7b1abdc9a43c17ed20a635ac6f93f441:

```
OLD (default serializer):

int store:      0.05s, 58M
int fetch:      0.25s, 58M
string store:   0.57s, 565M
string fetch:   1.01s, 565M
array store:    0.79s, 2169M
array fetch:    2.92s, 2169M
object store:   0.05s, 58M
object fetch:   0.58s, 58M

NEW (default serializer):

int store:      0.03s, 16M
int fetch:      0.16s, 16M
string store:   0.56s, 496M
string fetch:   0.93s, 496M
array store:    0.83s, 526M
array fetch:    1.32s, 526M
object store:   0.04s, 21M
object fetch:   0.44s, 21M

OLD (php serializer):

int store:      0.05s, 58M
int fetch:      0.25s, 58M
string store:   0.58s, 565M
string fetch:   1.03s, 565M
array store:    1.16s, 460M
array fetch:    14.90s, 460M
object store:   0.05s, 58M
object fetch:   0.59s, 58M

NEW (php serializer):

int store:      0.03s, 16M
int fetch:      0.17s, 16M
string store:   0.56s, 496M
string fetch:   0.95s, 496M
array store:    1.16s, 392M
array fetch:    14.33s, 392M
object store:   0.04s, 21M
object fetch:   0.44s, 21M
```