@staabm 2018-10-08
Just reading the numbers: it seems storing a array into apcu is waaaay slower then any other type.
Should we try to get arround storing a array from a userland perspective?