@nikic 2018-10-08
Due to a benchmark error I did not notice the regression in array persist/unpersist performance. With #324 the numbers there look better:

```
int store:      0.03s, 16M
int fetch:      0.17s, 16M
string store:   0.58s, 496M
string fetch:   0.96s, 496M
array store:    0.94s, 1549M
array fetch:    1.74s, 1549M
object store:   0.04s, 21M
object fetch:   0.46s, 21M
```

@staabm The benchmark script compares very different cases. "array" are arrays with sizes between 0 and 1000, while the objects are always empty `stdClass`es. Generally objects are the slowest type to store/fetch, because they will always use PHPs serialization mechanism (which is extremely slow). This is not intended as a comparison between different types.